+++ src/availables/gnoland.ts
@@ -1,0 +1,1 @@
+export { default } from "./testnet/gnoland";